Revert "SSCS-7398: CCD UI – When adding NINO then automatically associate cas… (#1223)"
This reverts commit b8992fbaaa6f91ed5269d9293cb7a59dbfac834e.
